Fix integer overflow in class total instance size.
This allows correct display and histogram sorting of the total instance
sizes that exceed 2G.

+++ src/com/sun/tools/hat/internal/model/JavaClass.java
@@ -382,7 +382,7 @@ public class JavaClass extends JavaHeapObject {
     public long getTotalInstanceSize() {
         int count = instances.size();
         if (count == 0 || !isArray()) {
-            return count * instanceSize;
+            return (long) count * instanceSize;
         }
 
         // array class and non-zero count, we have to
